requirements.txt
pandas
numpy
yfinance
matplotlib
python-dotenv
openai
